basic benchmark for new decimal type
